requirements.txt
# Paquetes y dependencias a utilizar en el proyecto

# Bot Telegram
python-telegram-bot>=20.0

# Azure
azure-search-documents>=11.4.0
azure-core>=1.30.0
azure-cosmos>=4.5.0
azure-monitor-opentelemetry>=1.0.0
opencensus-ext-azure>=1.1.0

# OpenAI
openai>=1.0.0

# Python
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
httpx>=0.25.0
pydantic-settings>=2.0.0